New Workshops: AIS and OpenSky with Grafana (#12)
What is included in this update?

Two new workshops have been developed to explore the features of mobilityDB while integrating it's capabilities with Grafana, and
open-source business intelligence dashboard platform.

Workshop Expansion (AIS Ship Data)

The first workshop, expands on the exist AIS Ship data and builds on existing queries from the workshop. The workshop covers:
- setting up a Grafana server
- connecting with Postgres
- completing basic SQL queries
- using MobilityDB specific functionality
- integrating Grafana variables (user-generated and global) with
MobilityDB
- developing dynamic queries based on user input

New Workshop (Open-Sky Flight Data)

The second new workshop, uses 24hrs of Open-Sky Network data and explores flight trajectories and airplane traffic. In addition to the above, this workshop includes:
- using multiple Grafana SQL queries in the same visualization
- best practice when adding Grafana global variables
- nested temporal filters from based on multiple parameters

Contribution Support for DocBook v4.5

This pull request also includes some minor adjustments to the previously included README.md file, and adds a CONTRIBUTING.md file with some preliminary guidance on how to convert a workshop document from Notion to DocBook using pandoc and dblatex so that it can be included in the compiled
workshop file. As of the writing of this document, the workshop
documents are generated using DocBook v4.5

How did we do it?

The work was completed collaboratively by:
Mohammad Ismail Tirmizi and Adam Broniewski with the support of Esteban Zimanyi and Mahmoud Sakr who both provided technical guidance and business insight during workshop development.

Former-commit-id: 655051b2fb676f1bd0bc3cbb79473ea34f246c90

Co-authored-by: Mohammad Ismail Tirmizi <ismailhero@gmail.com>
Co-authored-by: Adam Broniewski <abroniewski@gmail.com>